# api/nodejs-server/api/swagger.yaml
swagger: "2.0"
info:
  description: "Provides a RESTful API for interacting with Smartsplit. Documentation is written in both French and English."
  version: "v1"
  title: "Smartsplit API"
  x-origin:
  - format: "yaml"
    url: "https://raw.githubusercontent.com/iptoki/smartsplit/develop/api/swagger.yaml"
  x-providerName: "smartsplit.org"
  x-logo:
    url: "http://smartsplit.org/wp-content/uploads/2019/05/Logo-v02-005-noir.png"
#host: api.smartsplit.org
basePath: "/v1"
...
schemes:
- "https"
- "http"
consumes:
- "application/json"
produces:
- "application/json"
security:
- jwt: []
paths:
...
  /user:
    post:
      tags:
      - Users
      description: "Creates a new user account"
      parameters:
      - in: "body"
        name: "body"
        description: "Profile data and settings to be assigned to the new user"
        required: true
        schema:
          $ref: "#/definitions/user_signup"
      responses:
        200:
          description: "OK: User created"
          schema:
            $ref: "#/definitions/user_read"
        409:
          description: "Conflict: User already exists"
      x-swagger-router-controller: "Users"
      operationId: "createNewUser"
...
securityDefinitions:
  basicAuth:
    description: "HTTP Basic Authentication"
    type: "basic"
  jwt:
    description: "A JSON Web Token (JWT) used for access control and authorization"
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  rightHolder:
    type: "object"
    properties:
      rightHolderId:
        type: "string"
        description: "The unique Id for the right holder shared with AWS cognito user pools"
        example: "12345c60-7b1a-11e8-9c9c-2d42b21b1a4g"
      firstName:
        type: "string"
        example: "John"
        description: "The first name displayed on the right holder's profile"
      lastName:
        type: "string"
        example: "Smith"
        description: "The last name displayed on the right holder's profile"
      email:
        type: "string"
        example: "john.smith@example.com"
        description: "The email address for this right holder profile"
      requestSource:
        type: "string"
        example: "pochette"
        description: "The last known application that the user made a request to update account/profile"
      jurisdiction:
        type: "string"
        example: "Canada"
        description: "The country that has jurisdiction over the rights of the right holder"
      ipi:
        type: "string"
        example: "00004576"
        description: "The right holder's interested party information number"
      wallet:
        type: "string"
        example: "0xdd87ae15f4be97e2739c9069ddef674f907d27a8"
        description: "The blockchain wallet address for this right holder profile"
      avatarImage:
        type: "string"
        description: "The file name key in AWS S3 for the given right holder avatar image"
        example: "image.jpg"
      defaultRoles:
        type: "array"
        description: "The array of the default roles for the given right holder profile" 
        items:
          $ref: "#/definitions/rightHolder/properties/defaultRoles"
      groups:
        type: "array"
        description: "The array of the groups for the given right holder profile" 
        items:
          $ref: "#/definitions/rightHolder/properties/groups"
      artistName:
        type: "string"
        description: "The artist name for the given right Holder"
        example: "Questlove"
      accountCreationType: 
        type: "string"
        description: "The string value that describes if the account is self registered or create by a split initiator"
        example: "initiatorCreatedUser"
      locale: 
        type: "string"
        description: "The string value of the locale detected from the browser upon user or initiator account creation"
        example: "fr"
      socialMediaLinks:
        type: "object"
        description: "The urls for the social media profiles for the given right holder, like Facebook, Instagram, Youtube, Twitter"
        additionalProperties:
          type: "string"
          description: "social media links for the given right holder profile"
      editeur:
        type: "boolean"
        description: "Est un ayant-droit éditeur"        
    example:
      rightHolderId: "12345c60-7b1a-11e8-9c9c-2d42b21b1a4g"
      firstName: "John"
      lastName: "Smith"
      email: "john.smith@example.com"
      requestSource: "pochette"
      jurisdiction: "Canada"
      ipi: "00004576"
      wallet: "0xdd87ae15f4be97e2739c9069ddef674f907d27a8"
      avatarImage: "image.jpg"
      artistName: "Questlove"
      groups:
      - "Group 1"
      - "Group 2"
      defaultRoles:
      - "Songwriter"
      - "Producer"
      accountCreationType: "initiatorCreatedUser"
      locale: "en"
      editeur: true
      socialMediaLinks: 
        facebook: "https://facebook.com/ex"
        twitter: "https://twitter.com/ex"
        youtube: "https://youtube.com/ex"
...
  user_common:
    type: "object"
    properties:
      email:
        type: "string"
        format: "email"
        example: "maxime@iptoki.com"
        description: "The user's email address"
      requestSource:
        type: "string"
        example: "pochette"
        description: "The last known application that the user made a request to update account/profile"
      firstName:
        type: "string"
        example: "John"
        description: "The first name displayed on the right holder's profile"
      lastName:
        type: "string"
        example: "Smith"
        description: "The last name displayed on the right holder's profile"
      accountCreationType: 
        type: "string"
        description: "The string value that describes if the account is self registered or create by a split initiator"
        example: "initiatorCreatedUser"
      locale: 
        type: "string"
        description: "The string value of the locale detected from the browser upon user or initiator account creation"
        example: "fr"
    required:
      - email
  user_read:
    allOf:
    - $ref: "#/definitions/user_common"
    - type: "object"
      properties:
        userId:
          type: "string"
          example: "a1ed6d95-5342-4242-b0d1-99ad5786427e"
          description: "The user's unique UUID"
          format: "uuid"
        rightHolders:
          type: "array"
          description: "The UUID of the right holders this user has access to"
          example:
            - a1ed6d95-5342-4242-b0d1-99ad5786427e
          items:
            type: "string"
            format: "uuid"
  user_write:
    allOf:
    - $ref: "#/definitions/user_common"
    - type: "object"
      properties:
        password:
          type: "string"
          example: "correct horse battery staple"
          description: "The user's password"
  user_signup:
    allOf:
      - $ref: '#/definitions/user_common'
      - $ref: '#/definitions/user_write'
      - type: "object"
        properties:
          rightHolder:
            $ref: "#/definitions/rightHolder"